requirements.txt
python-dotenv==0.20.0
requests==2.27.1
pywin32==305